backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
httpx==0.28.1
Pillow>=12.0.0
PyMuPDF>=1.25.3
aiosqlite==0.20.0
pydantic>=2.10.0
pyjwt>=2.8.0
cryptography>=42.0.0
python-dotenv>=1.0.0
